% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{.get_files}
\alias{.get_files}
\title{read simulation result files from STICS workspace directory}
\usage{
.get_files(...)
}
\arguments{
\item{exp_dir}{(string) name of directory where experiment files are stored}

\item{exp_name_format}{(string) format of experiment name}

\item{usm_name_format}{(string) file name format}

\item{js_path}{(string) path to STICS javascript directory}

\item{ws}{(string) path to workspace}

\item{dir}{(string) name of directory where results are stored (exp_dir parent folder)}

\item{usm_list}{(char) list of usms names}

\item{ver_num}{(numeric)(optional) version number}

\item{stncode}{(string)(optional) weather station code id}

\item{soilcode}{(string)(optional) soil code id}

\item{ssp}{(string)(optional) ssp}

\item{group}{(string)(optional) variables to group by if returning list}

\item{return_type}{0 = list 1 = data.frame}
}
\value{
a data.table or a list of named data.tables of each usm.
}
\description{
`.get_files` returns a data.table or a list of data.tables of STICS simulations
}
