% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_files.R
\name{get_climate_files}
\alias{get_climate_files}
\title{get climate files}
\usage{
get_climate_files(
  name_format = "stn name",
  js_path = javastics_path,
  ws = workspace,
  return_type = 0
)
}
\arguments{
\item{name_format}{(string) file name format}

\item{js_path}{(string) path to STICS javascript directory}

\item{ws}{(string) path to workspace}

\item{return_type}{0 = list 1 = data.frame}
}
\description{
get climate files
}
