% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_s.R
\name{get_mod_s}
\alias{get_mod_s}
\title{get mod s files}
\usage{
get_mod_s(
  exp_dir,
  exp_name_format = "expid_runid",
  usm_name_format = "model_ssp_year_stncode_soilcode_verid",
  js_path = javastics_path,
  ws = workspace,
  dir = "RESULTS",
  usm_list = character(),
  ver_num = NULL,
  stn_code = character(),
  soil_code = character(),
  ssp = character(),
  group = NULL,
  return_type = 0
)
}
\arguments{
\item{exp_dir}{(string) name of directory where experiment files are stored}

\item{exp_name_format}{(string) format of experiment name}

\item{usm_name_format}{(string) file name format}

\item{js_path}{(string) path to STICS javascript directory}

\item{ws}{(string) path to workspace}

\item{dir}{(string) name of directory where results are stored (exp_dir parent folder)}

\item{usm_list}{(char) list of usms names}

\item{ver_num}{(numeric)(optional) version number}

\item{ssp}{(string)(optional) ssp}

\item{group}{(string)(optional) variables to group by if returning list}

\item{return_type}{0 = list 1 = data.frame}

\item{stncode}{(string)(optional) weather station code id}

\item{soilcode}{(string)(optional) soil code id}
}
\description{
get mod s files
}
