% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{.get_files}
\alias{.get_files}
\title{read simulation result files from STICS workspace directory}
\usage{
.get_files(...)
}
\value{
a data.table or a list of named data.tables of each usm.
}
\description{
`.get_files` returns a data.table or a list of data.tables of STICS simulations
}
